Fixed some byte string py3 compatibility
There's a spot where a string is being lengthened but the in python 3 it
is trying to add a byte string. The initialization of the string needs
the 'b' in front of it as such b''. Later the data that was pulled as a
byte string in py 3 is compared to a regular string and is failing the
check. I made the comparison string a byte string as well ('SunS' ->
b'Suns'). Python 2 ignores the b at the beginning of the strings and
treats them as regular string objects.