Changed the tests to make use of python standard library
The tests were hardcoded. I have changed them to run using the python
standard library's unittest module. The limitations of hardcoding made
themselves apparent when, after the switch, it says that three more
tests are being run than had been run before. This means that the
process of adding tests before was prone to missing some. In addition
there is now documentation on how to use the test suite that would not
have existed in the hardcoding case.