Add appveyor.yml for Windows testing/wheel/sdist
For this to get used:

* Create an AppVeyor account and add this GitHub repository to it
* Settings I choose to change
  * General
    * Build version format
      * {build}
      * Note that with vcversioner or similar this can be overwritten to include the calculated version
  * Environment
    * Environment variables
      * `APPVEYOR_RDP_PASSWORD`
        * Whatever password you might want.  You can click the lock icon to the right to mask the value.  This would only be accessible to people you give permission to to change settings.
      * `ENABLE_RDP`
        * With my code leaving it empty disables it, setting to `1` enables

The [RDP](https://en.wikipedia.org/wiki/Remote_Desktop_Protocol) feature allows you to connect in with a full Windows desktop during the build and check files and try building stuff.  This eases debugging of build issues.  Hopefully this wouldn't need to get used much.

The rest of the configuration is here in the `appveyor.yml` file.